[refactor] 감상 기록 / 패키지 수정 (#97)
