+++ android/app/src/main/java/com/teamtripdraw/android/ui/common/TripDrawViewModelFactory.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewmodel.CreationExtras
 import com.teamtripdraw.android.TripDrawApplication
 import com.teamtripdraw.android.ui.home.HomeViewModel
-import com.teamtripdraw.android.ui.post.writing.PostWritingViewModel
+import com.teamtripdraw.android.ui.postWriting.PostWritingViewModel
 import com.teamtripdraw.android.ui.signUp.NicknameSetupViewModel
 
 private const val UNDEFINED_VIEW_MODEL_ERROR = "ViewModelFactory에 정의되지않은 뷰모델을 생성하였습니다 : %s"
